Initial commit see message.
The initial is the first steps to layout a new project skeleton. We are using negroni and gorilla web packages to speed up development. The application has been broken down into the main application and a web context so that at a later stage we can build command line tools with out having to untangle the packages.  Im aiming to use the web context as a handler so the we don’t need  keep global context and/or use unsafe types. This still needs some development and refining. There was a lot of helpful information on the negroni project page. The configuration object is  not right yet. I focused a little to much on encapsulation and protecting the user of that lib and it could probably be simplified if those considerations weren’t taken into account. The yaml config object is also using json because the yaml libs I found didn’t have a way to make a custom unmarshaler. At frist I implemented the config object to hold all envs in memory by I didn’t like the way it felt so I guess im waiting for the raw yaml or on writing my own helpers.  For now it’s a one off read so speed is not critical.  I have the basics of the docker box that I will be using for development set up. I will probably be building the golang docker box based on the communities on build docker box in the next phase. Im still not clear what to do about the panic situation. I would love to have access logs even when there is a panic. Might need 2 layers or panic recovery. Guess thats a rabbit hole for another day.